rearranging tests

+++ cv1/src/main/java/vision/VisionUtil.java
@@ -87,9 +87,9 @@ public abstract class VisionUtil {
      * @return Rodrigues rotation vector
      */
     public static Mat panTilt(double pan, double tilt) {
-        Mat panV = Mat.zeros(3, 1, CvType.CV_32F);
+        Mat panV = Mat.zeros(3, 1, CvType.CV_64F);
         panV.put(0, 0, 0.0, pan, 0.0);
-        Mat tiltV = Mat.zeros(3, 1, CvType.CV_32F);
+        Mat tiltV = Mat.zeros(3, 1, CvType.CV_64F);
         tiltV.put(0, 0, tilt, 0.0, 0.0);
         return combineRotations(panV, tiltV);
     }
@@ -103,7 +103,7 @@ public abstract class VisionUtil {
      * @return camera intrinsic matrix
      */
     public static Mat makeIntrinsicMatrix(double f, Size dsize) {
-        Mat kMat = Mat.zeros(3, 3, CvType.CV_32F);
+        Mat kMat = Mat.zeros(3, 3, CvType.CV_64F);
         kMat.put(0, 0,
                 f, 0.0, dsize.width / 2,
                 0.0, f, dsize.height / 2,
@@ -447,7 +447,7 @@ public abstract class VisionUtil {
             MatOfDouble dMat,
             MatOfPoint3f targetGeometryMeters) {
         // System.out.println(targetGeometryMeters.dump());
-        Mat worldTVec = Mat.zeros(3, 1, CvType.CV_32F);
+        Mat worldTVec = Mat.zeros(3, 1, CvType.CV_64F);
         worldTVec.put(0, 0, xPos, yPos, zPos);
 
         // camera up/right means world down/left, so both negative
